deployment/requirements.txt
gradio
tensorflow
opencv-python
scikit-learn
matplotlib
numpy
pillow